Trace and Role
* fixed trace API to add directly new records
* changed role namespace from '::Record::' to '::'
