+++ src/Record.h
@@ -45,6 +45,7 @@
 
 #include <libtptp/Formula>
 #include <libtptp/Node>
+#include <libtptp/Role>
 
 /**
    @brief    TODO
@@ -61,21 +62,6 @@ namespace libtptp
     {
       public:
         using Ptr = std::shared_ptr< Record >;
-
-        enum class Role
-        {
-            AXIOM,              //!<  TODO: PPA: FIXME: description
-            HYPOTHESIS,         //!<  TODO: PPA: FIXME: description
-            DEFINITION,         //!<  TODO: PPA: FIXME: description
-            ASSUMPTION,         //!<  TODO: PPA: FIXME: description
-            LEMMA,              //!<  TODO: PPA: FIXME: description
-            THEOREM,            //!<  TODO: PPA: FIXME: description
-            CONJECTURE,         //!<  TODO: PPA: FIXME: description
-            NEGATED_CONJECTURE, //!<  TODO: PPA: FIXME: description
-            PLAIN,              //!<  TODO: PPA: FIXME: description
-            TYPE,               //!<  TODO: PPA: FIXME: description
-            UNKNOWN,            //!<  TODO: PPA: FIXME: description
-        };
 
         Record( const Identifier::Ptr& identifier, const Role role,
             const Formula::Ptr& formula );
